requirements.txt
openenv-core
fastapi
uvicorn
pydantic